0.0.2

+++ src/controllers/todo.js
@@ -3,15 +3,23 @@ import responseStructure from "../../utils/responseStructure.js";
 
 export const add_todo = async (data, cb) => {
   try {
-    if (!data.user_id || !data.todo || !data.start || !data.end)
+    console.log(data);
+    if (
+      !data.todo ||
+      !data.start ||
+      !data.end ||
+      !data.created_for ||
+      data.created_for.length <= 0
+    )
       throw new Error("Params missing");
 
     const newTodo = new todo({
-      user_id: data.user_id,
+      created_by: data.user._id,
       todo: data.todo,
       details: data.details || "",
       start: new Date(data.start),
       end: new Date(data.end),
+      created_for: data.created_for,
     });
 
     const res = await newTodo.save();
@@ -43,21 +51,24 @@ export const add_todo = async (data, cb) => {
 
 export const get_todo = async (data, cb) => {
   try {
-    const where_data = {};
+    const where_data = {
+      created_by: data.user._id,
+    };
+
     if (data.date) {
-      where_data.created_at = new Date(data.date);
+      where_data.createdAt = new Date(data.date);
     } else if (data.start && data.end) {
-      where_data.created_at = {
+      where_data.createdAt = {
         $gte: new Date(data.start),
         $lte: new Date(data.end),
       };
     } else {
       const latest_date = await todo.findOne(
         { user_id: data.user_id },
-        { created_at: true },
-        { sort: { created_at: -1 } }
+        { createdAt: true },
+        { sort: { createdAt: -1 } }
       );
-      where_data.created_at = latest_date?.created_at;
+      where_data.createdAt = latest_date?.createdAt;
     }
 
     const res = await todo.find(where_data);
@@ -91,6 +102,13 @@ export const delete_todo = async (data, cb) => {
   try {
     if (!data.id) throw new Error("Params missing");
 
+    const found_todo = await todo.findById(data.id);
+    if (!found_todo) throw new Error("no todo found");
+
+    if (found_todo.created_by !== data.user._id) {
+      throw new Error("You dont have permission to delete this todo");
+    }
+
     const res = await todo.findByIdAndDelete(data.id);
 
     return cb(
@@ -121,6 +139,17 @@ export const delete_todo = async (data, cb) => {
 export const update_todo = async (data, cb) => {
   try {
     if (!data.id) throw new Error("Params missing");
+
+    const found_todo = await todo.findById(data.id);
+    if (!found_todo) throw new Error("no todo found");
+
+    if (
+      found_todo.created_by !== data.user._id ||
+      !found_todo.created_for.includes(data.user._id)
+    ) {
+      throw new Error("You dont have permission to update this todo");
+    }
+
     const updated_data = {};
     if (data.todo) {
       updated_data.todo = data.todo;
@@ -134,66 +163,13 @@ export const update_todo = async (data, cb) => {
     if (data.end) {
       updated_data.end = new Date(data.end);
     }
+    if (data.status) {
+      updated_data.status = data.status;
+    }
 
     updated_data.updated_at = new Date();
 
     const res = await todo.findByIdAndUpdate(data.id, updated_data);
-
-    return cb(
-      null,
-      responseStructure
-        .merge({
-          status: 200,
-          success: true,
-          data: res,
-          message: "ok",
-        })
-        .toJS()
-    );
-  } catch (err) {
-    return cb(
-      responseStructure
-        .merge({
-          status: 400,
-          success: false,
-          data: null,
-          message: err.message,
-        })
-        .toJS()
-    );
-  }
-};
-
-export const update_todo_status = async (data, cb) => {
-  try {
-    if (!data.id) throw new Error("Params missing");
-
-    const found = await todo.findOne({ _id: data.id });
-    if (!found) throw new Error("todo not found");
-
-    if (found.is_done) {
-      throw new Error("This todo is alreday done, you cant change its status");
-    }
-
-    if (data.is_active === found.is_active) {
-      throw new Error(
-        `this todo is already marked as ${
-          data.is_active ? "active" : "inative"
-        }`
-      );
-    }
-
-    const updated_data = {};
-    if (data.is_active === true || data.is_active === false) {
-      updated_data.is_active = data.is_active;
-    } else if (data.is_done === true) {
-      updated_data.is_active = false;
-      updated_data.is_done = true;
-    }
-
-    const res = await todo.findByIdAndUpdate(data.id, updated_data, {
-      new: true,
-    });
 
     return cb(
       null,
